Adding Angular modules and optimizing angular apps
Videos with Max. I liked being able to add more modules in order to condense the amount of code that was in some of the files. It is easier to read and easier to navigate as well. I feel so confident when I code along with him, I just hope that with time I feel more confident in my self.